+++ exonum/__init__.py
@@ -1,1 +1,5 @@
 # coding: utf-8
+
+from .client import ExonumClient
+from .message import MessageGenerator, ExonumMessage, gen_keypair
+from .module_manager import ModuleManager