+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### Unreleased
+
+- The client has been adopted to work with Exonum dynamic_services branch.
+- Proto files are now downloaded via REST API and compiled dynamically.
+- Tx generation is now protobuf-based.
+
 ### 0.2 - 2019-03-29
 
 - The client has been adopted to work with Exonum 0.10.x.